pkmn-classifier/requirements.txt
torch
torchvision
transformers
Pillow
kafka-python
requests
